--- server/__tests__/app.test.js
@@ -1,7 +1,0 @@
-const app = require('../app');
-
-describe('app', () => {
-  it('instantiates properly', () => {
-    expect(app).toBeDefined();
-  });
-});
