+++ server/__tests__/initApp.test.js
@@ -1,0 +1,7 @@
+const initApp = require('../initApp');
+
+describe('initApp', () => {
+  it('returns a function', () => {
+    expect(initApp()).toBeInstanceOf(Function)
+  });
+});
