+++ server/server.js
@@ -1,8 +1,9 @@
-const app = require('./app');
+const initApp = require('./initApp');
 const mongoDB = require('./database/mongoDB');
 
-const port = process.env.PORT || 5000;
+const app = initApp();
 
 mongoDB.connect();
 
+const port = process.env.PORT || 5000;
 app.listen(port, console.log.bind(null, `Listening on port ${port}`));
